requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distro==1.8.0
Flask==3.0.0
fonttools==4.43.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mido==1.3.0
numpy==1.26.1
openai==1.1.1
packaging==23.2
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.1
